requirements.txt
Django==1.11.12
gcloud==0.17.0
googleapis-common-protos==1.5.3
httplib2==0.11.0
jws==0.1.3
oauth2client==3.0.0
pbr==3.1.1
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2018.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.11.0
stevedore==1.28.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
django-audit-log